게시판 비밀번호 기능 추가

+++ src/main/java/com/example/miniproj2/contoller/BoardController.java
@@ -2,14 +2,17 @@ package com.example.miniproj2.contoller;
 
 import com.example.miniproj2.entity.Board;
 import com.example.miniproj2.repository.BoardRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.support.RequestContextUtils;
 
 import java.util.List;
+import java.util.Map;
 
 @Controller
 @RequestMapping("/board")
@@ -59,7 +62,13 @@ public class BoardController {
     }
 
     @GetMapping("/edit")
-    public String edit(@RequestParam long id, Model model){
+    public String edit(HttpServletRequest request, Model model){
+
+        Map<String, ?> map = RequestContextUtils.getInputFlashMap(request);
+        if(map == null) throw new RuntimeException("권한 없음");
+
+        Long id = (Long)map.get("id");
+
         Board board = boardRepository.findById(id).orElseThrow();
 
         model.addAttribute("board", board);
@@ -72,12 +81,15 @@ public class BoardController {
             @RequestParam("id") Long id,
             @RequestParam("title") String title,
             @RequestParam("writer") String writer,
-            @RequestParam("content") String content){
+            @RequestParam("content") String content,
+            @RequestParam("password") String password
+    ){
 
         Board origin = boardRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("id가 존재하지 않습니다."));
         origin.setTitle(title);
         origin.setWriter(writer);
         origin.setContent(content);
+        origin.setPassword(password);
 
         boardRepository.save(origin);
 
@@ -86,9 +98,38 @@ public class BoardController {
     }
 
     @GetMapping("/delete")
-    public String delete(@RequestParam("id") long id){
+    public String delete(HttpServletRequest request){
+
+        Map<String, ?> map = RequestContextUtils.getInputFlashMap(request);
+        if(map == null) throw new RuntimeException("권한 없음");
+
+        Long id = (Long)map.get("id");
+
+
         boardRepository.deleteById(id);
 
         return "redirect:/board/list";
+    }
+
+    @GetMapping("/password/{mode:edit|delete}/{id}")
+    public String password(@PathVariable String mode, @PathVariable("id") long id, Model model){
+        model.addAttribute("mode", mode);
+        model.addAttribute("id", id);
+
+        return "/board/password";
+    }
+
+    @PostMapping("/password/{mode:edit|delete}/{id}")
+    public String password(@PathVariable String mode, @PathVariable long id, @RequestParam String password, RedirectAttributes attr){
+        // 글번호가 없으면 오류 처리
+        Board board = boardRepository.findById(id).orElseThrow();
+
+        if(board.getPassword().equals(password)){
+            attr.addFlashAttribute("id", id);
+            return "redirect:/board/" + mode;
+        }
+        else {
+            return "redirect:/board/password/" + mode + "/" + id + "?error";
+        }
     }
 }
